moving decompression logic to its own module
